refactor : 폴더 구조 재정비 features 폴더 생성 및 파일 위치 변경 ref : #5
