+++ src/features/faq/ui/FAQButton.tsx
